+++ autoload.rb
@@ -3,5 +3,5 @@ require 'yaml'
 require_relative 'config/i18n'
 require_relative 'app/modules/in_out'
 require_relative 'app/modules/uploader'
-require_relative 'console'
-require_relative 'account'
+require_relative 'app/entities/console'
+require_relative 'app/entities/account'
